Find the Inno Setup compiler in the standard path
This patch implements a new property `compiler` to allow a user to specify
a single relative filename for the Inno Setup compiler rather than a wine
and windows compiler. Whether specified as `windows compiler`, `wine compiler`
or just `compiler` if the file does not exist an attempt to resolve it will be
made. The patch also allows for use of `ISCC.exe` instead of `Compil32.exe`
for quieter builds.